Make the parsing of various syntaxes less order-dependent
The old way to parse the buffer did not let the various
implementations easily control what other things could apply
afterwards. This led to undesirable behaviors, e.g. processing of macros
inside <pre> blocks.

Reorganized the code to have a weblog-process-regexp-operations
function that takes a list of (regexp implementation) pairs and
processes them in such a way that the first regexp that matches is
selected. Each implementation can then decide how it moves the
cursor. By default, the cursor stays at the current point, but an
implementation can move it at end of the modified section to prevent
further processing.

This also involved reworking the handling of <p></p> generation. I
beliveve that the new method is slightly better and faster.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>